refactor bucket picking strategies into own modules and clean up
