+++ app/assets/javascripts/oxalis/controller/viewmodes/arbitrary_controller.js
@@ -82,12 +82,6 @@ class ArbitraryController extends React.PureComponent<Props> {
     this.start();
   }
 
-  componentDidUpdate(prevProps: Props) {
-    if (prevProps.viewMode !== this.props.viewMode) {
-      this.plane.setMode(this.props.viewMode);
-    }
-  }
-
   componentWillUnmount() {
     this.stop();
   }
@@ -296,7 +290,6 @@ class ArbitraryController extends React.PureComponent<Props> {
             displayCrosshair,
             crosshairSize,
           } = userConfiguration;
-          this.plane.setSphericalCapRadius(sphericalCapRadius);
           this.setClippingDistance(clippingDistanceArbitrary);
           this.crosshair.setScale(crosshairSize);
           this.crosshair.setVisibility(displayCrosshair);
@@ -321,7 +314,6 @@ class ArbitraryController extends React.PureComponent<Props> {
     this.arbitraryView.start();
 
     this.plane = new ArbitraryPlane();
-    this.plane.setMode(this.props.viewMode);
     this.crosshair = new Crosshair(Store.getState().userConfiguration.crosshairSize);
     this.crosshair.setVisibility(Store.getState().userConfiguration.displayCrosshair);
